Struct advent::Solution

pub struct Solution<'a, I> { /* private fields */ }

A runner and benchmarker for an Advent of Code solution.

Constructed using advent::new.

Implementations

impl<'a, I> Solution<'a, I>where I: Clone + UnwindSafe,

pub fn run(self) -> Summary

Consumes this struct and runs the parts.

pub fn bench(self) -> Summary

Consumes this struct and benchmarks the parts.

pub fn cli(self)

Parses the command line arguments and executes the run or benchmark.