Crate advent

This crate contains a runner and benchmarker for Advent of Code solutions with free Christmas trees 🎄.

Getting started

Add this crate to your Cargo manifest for your solution.

advent = { git = "https://github.com/rossmacarthur/advent" }

Then use the following as your main function.

fn main() {
    let solution = advent::new(parse_input).part(part1).part(part2).build();
    solution.cli()
}

Where

parse_input is a function that returns any type I implementing Clone.

Each part function takes I as an argument and returns something implementing Display.

Finally, cli() will instantiate a command line interface and run the program. Ordinary runs will run each part once and output the answers. Passing --bench to the program will perform a benchmark.

✨ That’s all! You’re free to structure your program however else you want.

Features

There are also some optional features which pull in some other crates.

festive enables some festive ascii art and changes the default output to --output festive

json supports JSON output using --output json, useful for collecting benchmark information

prelude re-exports my prelude crate that can be imported using

use advent::prelude::*;

They can be enabled in your Cargo manifest like this:

[dependencies]
advent = { git = "https://github.com/rossmacarthur/advent", features = ["festive", "json"] }

CLI

The command line interface looks like this.

Usage: example [--bench] [--output <output>]

Run the program.

Options:
  --bench           whether to benchmark
  --output          the output style
  --help            display usage information

The output style can be boring, festive, or json. To use json this crate requires the json feature to be set.

Modules

summary

Structs

Builder

A builder for a Solution.

Solution

A runner and benchmarker for an Advent of Code solution.

Functions

new

Returns a new builder for a new Advent of Code run or benchmark using the given parse function.